--- GarcissNetLibs/02_Test/Common/Garciss.Core.Common.TestHelper/Formatters/TestOfuscacion.cs
@@ -1,19 +1,0 @@
-﻿using Garciss.Core.Common.Formatters;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace Garciss.Core.Common.TestHelper.Formatters {
-    [TestClass]
-    public class TestOfuscacion {
-        [TestMethod]
-        public void TestOfuscarCorreo() {
-            var ofuscando = Obfuscation.OfuscarCorreo("prueba@gmail.com");
-            Assert.AreEqual("****ba@gmail.com", ofuscando);
-        }
-
-        [TestMethod]
-        public void TestOfuscarTelefonoMovil() {
-            var ofuscando = Obfuscation.OfuscarMovil("666555777");
-            Assert.AreEqual("666*****7", ofuscando);
-        }
-    }
-}
