--- GarcissNetLibs/02_Test/Common/Garciss.Core.Common.TestHelper/Garciss.Core.Common.TestHelper.csproj
@@ -1,30 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.3" />
-    <PackageReference Include="MSTest.TestAdapter" Version="2.1.2" />
-    <PackageReference Include="MSTest.TestFramework" Version="2.1.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\01_GarcissNetLibs\Business\Garciss.Core.Business.Importe\Garciss.Core.Business.Importe.csproj" />
-    <ProjectReference Include="..\..\..\01_GarcissNetLibs\Common\Garciss.Core.Common.Converter\Garciss.Core.Common.Converter.csproj" />
-    <ProjectReference Include="..\..\..\01_GarcissNetLibs\Common\Garciss.Core.Common.EnumHelper\Garciss.Core.Common.EnumHelper.csproj" />
-    <ProjectReference Include="..\..\..\01_GarcissNetLibs\Common\Garciss.Core.Common.Extensions\Garciss.Core.Common.Extensions.csproj" />
-    <ProjectReference Include="..\..\..\01_GarcissNetLibs\Common\Garciss.Core.Common.Formatters\Garciss.Core.Common.Formatters.csproj" />
-    <ProjectReference Include="..\..\..\01_GarcissNetLibs\Common\Garciss.Core.Common.Helper\Garciss.Core.Common.Helper.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="Converters\Fake\TestXml.xml">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-</Project>
