--- GarcissNetLibs/02_Test/Common/Garciss.Core.Common.TestRespuesta/Garciss.Core.Common.TestRespuesta.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.3" />
-    <PackageReference Include="MSTest.TestAdapter" Version="2.1.2" />
-    <PackageReference Include="MSTest.TestFramework" Version="2.1.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\01_GarcissNetLibs\Common\Garciss.Core.Common.Respuestas\Garciss.Core.Common.Respuestas.csproj" />
-  </ItemGroup>
-
-</Project>
