--- GarcissNetLibs/02_Test/Common/Garciss.Core.Common.TestSession/UnitTest1.cs
@@ -1,10 +1,0 @@
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace Garciss.Core.Common.TestSession {
-    [TestClass]
-    public class UnitTest1 {
-        [TestMethod]
-        public void TestMethod1() {
-        }
-    }
-}
