+++ GarcissNetLibs/02_Test/Garciss.Testing.MoqTestingData/Data/Email/FakeData/FakeDataReturnOkData.cs
@@ -1,0 +1,11 @@
+﻿using Garciss.Core.Common.Respuestas;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Garciss.Testing.MoqTestingData.Data.Email.FakeData {
+    internal static class FakeDataReturnOkData {
+    }
+}
