+++ GarcissNetLibs/02_Test/Garciss.Testing.MoqTestingData/Garciss.Testing.MoqTestingData.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Moq" Version="4.16.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\01_GarcissNetLibs\Data\Garciss.Core.Data.Email\Garciss.Core.Data.Email.csproj" />
+  </ItemGroup>
+
+</Project>
