--- GarcissNetLibs/02_Test/Libs/Garciss.Core.Libs.TestEncriptacion/Garciss.Core.Libs.TestEncriptacion.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.3" />
-    <PackageReference Include="MSTest.TestAdapter" Version="2.1.2" />
-    <PackageReference Include="MSTest.TestFramework" Version="2.1.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\01_GarcissNetLibs\Libs\Garciss.Core.Libs.Encriptacion.AES\Garciss.Core.Libs.Encriptacion.AES.csproj" />
-  </ItemGroup>
-
-</Project>
