+++ GarcissNetLibs/02_Test/UnitTest/Common/Garciss.Core.Common.TestHelper/Converters/Fake/ClaseParaDeserializar.cs
