+++ GarcissNetLibs/02_Test/UnitTest/Common/Garciss.Core.Common.TestHelper/EnumHelper/EnumTestFake.cs
@@ -1,0 +1,23 @@
+﻿using Garciss.Core.Common.Helper.EnumHelper;
+
+namespace Garciss.Core.Common.TestHelper.EnumHelper {
+    public enum EnumTestFake {
+        unknow = -1,
+        [EnumDescription(@"00")]
+        testNumCero = 0,
+        [EnumDescription(@"10")]
+        testNumUno = 1,
+        [EnumDescription(@"Dos")]
+        testNumDos = 2,
+        [EnumDescription(@"hgfjrhguor")]
+        testNumDiez = 3,
+    }
+
+    public enum EnumTestFakeSinClaves {
+        desconocido = -1,
+        enumSinClave0 = 0,
+        enumSinClave1 = 1,
+        enumSinClave2 = 2,
+        enumSinClave3 = 3
+    }
+}
