+++ GarcissNetLibs/02_Test/UnitTest/Common/Garciss.Core.Common.TestSession/UnitTest1.cs
@@ -1,0 +1,13 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Garciss.Core.Common.TestSession
+{
+    [TestClass]
+    public class UnitTest1
+    {
+        [TestMethod]
+        public void TestMethod1()
+        {
+        }
+    }
+}
