+++ GarcissNetLibs/02_Test/UnitTest/Data/Garciss.Core.Data.TestEmail/App.config
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <system.net>
+    <mailSettings>
+      <smtp deliveryMethod="SpecifiedPickupDirectory">
+        <specifiedPickupDirectory
+          pickupDirectoryLocation="c:\maildrop"/>
+      </smtp>
+    </mailSettings>
+  </system.net>
+</configuration>
