+++ GarcissNetLibs/02_Test/UnitTest/Data/Garciss.Core.Data.TestEmail/Garciss.Core.Data.TestEmail.csproj
@@ -1,0 +1,27 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+
+    <IsPackable>false</IsPackable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.3" />
+    <PackageReference Include="MSTest.TestAdapter" Version="2.1.2" />
+    <PackageReference Include="MSTest.TestFramework" Version="2.1.2" />
+    <PackageReference Include="coverlet.collector" Version="1.3.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\..\01_GarcissNetLibs\Data\Garciss.Core.Data.Email\Garciss.Core.Data.Email.csproj" />
+    <ProjectReference Include="..\..\..\Garciss.Testing.MoqTestingData\Garciss.Testing.MoqTestingData.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="Plantillas\EmailResponsiveTemplate.html">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
+
+</Project>
