+++ GarcissNetLibs/02_Test/UnitTest/Data/Garciss.Core.Data.TestFiles/TestArchivos.cs
@@ -1,0 +1,18 @@
+﻿using Garciss.Core.Data.Files;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System.IO;
+
+namespace Core.Data.TestFiles {
+    [TestClass]
+    public class TestArchivos {
+
+        [TestMethod]
+        public void Ordenar() {
+            var archivos = Directory.GetDirectories("prueba", "", SearchOption.AllDirectories);
+
+            var listaOrdenada = Archivos.FicherosOrdenados(archivos);
+
+            Assert.IsTrue(listaOrdenada.Count == archivos.Length);
+        }
+    }
+}
