+++ GarcissNetLibs/02_Test/UnitTest/Data/Garciss.Core.Data.TestFiles/TestDirectorios.cs
@@ -1,0 +1,40 @@
+﻿using Garciss.Core.Data.Files;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+
+namespace Core.Data.TestFiles {
+    [TestClass]
+    public class TestDirectorios {
+
+        [TestMethod]
+        public void Copiar() {
+            try {
+                Directorios.Copy(
+                    new DirectoryInfo("prueba"),
+                    new DirectoryInfo("copia")
+                );
+
+                var archivosOrigen = new List<FileInfo>(new DirectoryInfo("prueba").GetFiles("*", SearchOption.AllDirectories));
+                var nombresOrigen = (from arch in archivosOrigen
+                                     select arch.Name).ToList();
+
+                var archivosDestino = new List<FileInfo>(new DirectoryInfo("copia").GetFiles("*", SearchOption.AllDirectories));
+                var nombresDestino = (from dest in archivosDestino
+                                      select dest.Name).ToList();
+                archivosOrigen = null;
+                archivosDestino = null;
+
+                Assert.IsTrue(Directory.Exists("prueba") && Directory.Exists("copia"));
+                Assert.IsTrue(nombresOrigen.Count == nombresDestino.Count);
+
+                foreach (var origen in nombresOrigen)
+                    Assert.IsTrue(nombresOrigen.Contains(origen));
+
+            } finally {
+                Directory.Delete("copia", true);
+            }
+        }
+    }
+}
