+++ GarcissNetLibs/02_Test/UnitTest/Libs/Garciss.Core.Libs.TestEncriptacion/Garciss.Core.Libs.TestEncriptacion.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+
+    <IsPackable>false</IsPackable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.3" />
+    <PackageReference Include="MSTest.TestAdapter" Version="2.1.2" />
+    <PackageReference Include="MSTest.TestFramework" Version="2.1.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\..\01_GarcissNetLibs\Libs\Garciss.Core.Libs.Encriptacion\Garciss.Core.Libs.Encriptacion.csproj" />
+  </ItemGroup>
+
+</Project>
